refactor: move provider/session files into src/editor/
