+++ src/editor/hexEditorProvider.ts
